Elixir Diary v0.1.0 Diary.CLI

Outputs the diary notes on the CLI

Functions

create_note()
(since 0.1.0)

Specs

create_note() :: :ok

Gets the input of the note value and creates a note

greeting()
(since 0.1.0)

Specs

greeting() :: :ok

Outputs a greeting to the user

main(_)
(since 0.1.0)

Main function resposible to control the execution of the CLI logic

menu()
(since 0.1.0)

Specs

menu() :: :ok

Outputs a menu to the user

read_notes()
(since 0.1.0)

Specs

read_notes() :: [any()]

Loads the notes from the filesystem and outputs it to the user